Check if ro.boot.vhal_proxy_server_port is set
Because kBootConfigPort is cast to unsigned int type, the check for
ro.boot.vhal_proxy_server_port would always pass even if it's unset.

Let's fix this.

Bug: 427249685
Test: Verify the check fails when ro.boot.vhal_proxy_server_port is not set

+++ guest/hals/vehicle/VehicleService.cpp
@@ -65,15 +65,14 @@ int main(int argc, char* argv[]) {
   }
   ABinderProcess_startThreadPool();
 
-  unsigned int port =
-      static_cast<unsigned int>(property_get_int32(kBootConfigPort, -1));
+  int port = property_get_int32(kBootConfigPort, -1);
   CHECK(port >= 0) << "Failed to read port number from: " << kBootConfigPort;
 
   std::string serverAddr;
   if (useVsock) {
     VsockConnectionInfo vsock = {
         .cid = VMADDR_CID_HOST,
-        .port = port,
+        .port = static_cast<unsigned int>(port),
     };
     serverAddr = vsock.str();
     LOG(INFO) << "Connecting to vsock server at " << serverAddr;
